+++ CHANGELOG.md
@@ -14,11 +14,12 @@ All user visible changes to `cucumber` crate will be documented in this file. Th
 ### Added
 
 - Ability for step functions to return `Result`. ([#151])
-- `writer::JUnit` behind the `output-junit` feature flag ([#147])
 - Arbitrary output for `writer::Basic` ([#147])
+- `writer::JUnit` ([JUnit XML report][0110-1]) behind the `output-junit` feature flag ([#147])
 
 [#147]: /../../pull/147
 [#151]: /../../pull/151
+[0110-1]: https://llg.cubic.org/docs/junit